{{cookiecutter.repo_name}}/pyproject.toml
[tool.poetry]
name = "{{cookiecutter.project_name}}"
version = "0.1.0"
description = "An awesome Django backend for '{{cookiecutter.project_name}}' project"
authors = []

[tool.poetry.dependencies]
python = "^3.8"
Django = "^3.1"
django-environ = "^0.4.5"
whitenoise = "^5.2.0"
gunicorn = "^20.1.0"
psycopg2-binary = "^2.8.6"
djangorestframework = "^3.12.4"
django-rest-authemail = { git = "https://github.com/viper-development/django-rest-authemail.git", rev = "bb8dee2665153a769fa78bfbc422732d1843c636" }

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
